requirements.txt
fastapi[standard]
uvicorn
redis[async]
typer
python-dotenv
httpx
pydantic
